requirements.txt
fastapi
uvicorn[standard]
pydantic>=2,<3
python-docx
docxtpl
openai
